refactor: share material label dialog
