+++ Saber3/src/views/mes/mops/prodPlan/list.vue
@@ -111,7 +111,7 @@
 <script>
 import listPage from '@/mixins/list-page';
 import options from './list';
-import materialLabelDialog from './../solutionPlan/cpns/materialLabelDialog.vue';
+import materialLabelDialog from '../cpns/materialLabelDialog.vue';
 
 export default {
   mixins: [listPage],
